Rearrange CSS rules and update table cell selectors
This commit moves some margin-related CSS rules in `tailwind.css` for better structure and readability. Also, the `.fj-gap-10` is renamed to `.fj-gap-x-10` and is now placed in an appropriate position. In the `theme-overrides.css` file, selectors for table cells have been updated to only target `td` and `th` without a class, hence improving specificity.