Add Simple Stats module and related CSS changes
A new Simple Stats module has been created with associated fields and metadata defined in the JSON files. This module is designed to display statistical information with specified layout based on the number of stats items. Corresponding CSS changes have been introduced to support the layout and design aspects of the stats.