Add 'Stats - Simple Grid' module
The commit adds a new module, 'Stats - Simple Grid'. It consists of three files: fields.json, meta.json, and module.html. The module displays different types of stats in a simple grid layout, and it can be used in various content types including LANDING_PAGE, SITE_PAGE, BLOG_LISTING and BLOG_POST. The contents of the grid (title and subtext) and the number of grid columns can be customized.